+++ lib/quectocolors.dart
@@ -1,2 +1,6 @@
 export "supports_ansi_color.dart";
 export "src/quectocolors.dart"; // QuectoColors, Quecto
+export "src/quectocolors_extras.dart";  // QuectoColors, QuectoStyler, QuectoPlain
+// Adds String extensions for all the CSS/X11 color names.
+export "src/quectocolors_css.g.dart";
+