+++ lib/quectocolors_css.dart
@@ -1,5 +1,0 @@
-export "supports_ansi_color.dart";
-export "src/quectocolors.dart";  // QuectoColors, QuectoStyler, QuectoPlain
-// Adds String extensions for all the CSS/X11 color names.
-export "src/quectocolors_css.g.dart";
-