+++ lib/small.dart
@@ -1,0 +1,2 @@
+export "supports_ansi_color.dart";
+export "src/quectocolors.dart"; // QuectoColors, Quecto